use the same library reconciliation when importing via file picker
